pyproject.toml
[tool.poetry]
name = "peopen"
version = "0.1.0"
description = ""
authors = ["hazn <aitakprojects@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.3.0"
cookiecutter = "^2.6.0"
marimo = "^0.4.6"
numpy = "^1.26.4"
dspy-ai = "^2.4.5"
nltk = "^3.8.1"
rouge = "^1.0.1"
evaluate = "^0.4.1"
tiktoken = "^0.6.0"
bert-score = "^0.3.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
